requirements.txt
python_version >= '2.7'
tensorflow-gpu>=1.2.0
keras>=2.0.4
pillow-simd>=4.1.1
joblib>=0.11
h5py>=2.7.0
cython>=0.25.2
typing>=3.6.1
-e git://github.com/raghakot/keras-vis.git#egg=keras-vis
-e git+https://github.com/lucasb-eyer/pydensecrf.git
-e git+git://github.com/keplr-io/quiver.git
